+++ Binance-Spot-API/Model/Account/AccountTradeList.cs
@@ -13,13 +13,13 @@ namespace Binance_Spot_API.Model.Account
         [JsonProperty("orderListId")]
         public int OrderListId { get; set; }
         [JsonProperty("price")]
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         [JsonProperty("qty")]
-        public decimal Quantity { get; set; }
+        public double Quantity { get; set; }
         [JsonProperty("quoteQty")]
-        public decimal QuoteQuantity { get; set; }
+        public double QuoteQuantity { get; set; }
         [JsonProperty("commission")]
-        public decimal Commission { get; set; }
+        public double Commission { get; set; }
         [JsonProperty("commissionAsset")]
         public string CommissionAsset { get; set; }
         [JsonProperty("time")]
